+++ Documentation/CHANGELOG.md
@@ -1,4 +1,5 @@
-#Version control history:
+Version control history:
+====================
 
 * Version 1.0-1 020717
 	* First version
@@ -10,5 +11,8 @@
 	* GPIO LED feature added.
 
 * Version 1.3-4 140817
-	* Graph function added.
+	* Graph of logfile function added.
 	
+* Version 1.4-5 200817
+	* Code rewritten complete in python3 (from a mixture of bash and python)
+	* Graph of realtime CPU data added